+++ src/main/java/com/yourmode/yourmodebackend/domain/user/service/UserProfileServiceImpl.java
@@ -13,7 +13,6 @@ import com.yourmode.yourmodebackend.global.common.exception.RestApiException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import com.yourmode.yourmodebackend.domain.user.dto.request.PasswordUpdateRequestDto;
 import com.yourmode.yourmodebackend.domain.user.entity.UserCredential;
 import com.yourmode.yourmodebackend.domain.user.repository.UserCredentialRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -35,7 +34,7 @@ public class UserProfileServiceImpl implements UserProfileService {
      * @throws RestApiException USER_NOT_FOUND, PROFILE_NOT_FOUND 등
      */
     @Transactional(readOnly = true)
-    public UserProfileResponseDto getMyProfile(Long userId) {
+    public UserProfileResponseDto getMyProfile(Integer userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RestApiException(UserErrorStatus.USER_NOT_FOUND));
 
@@ -60,7 +59,7 @@ public class UserProfileServiceImpl implements UserProfileService {
      * @throws RestApiException USER_NOT_FOUND, PROFILE_NOT_FOUND, BODY_TYPE_NOT_FOUND 등
      */
     @Transactional
-    public UserProfileResponseDto updateMyProfile(Long userId, UserProfileUpdateRequestDto dto) {
+    public UserProfileResponseDto updateMyProfile(Integer userId, UserProfileUpdateRequestDto dto) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RestApiException(UserErrorStatus.USER_NOT_FOUND));
 
@@ -95,7 +94,7 @@ public class UserProfileServiceImpl implements UserProfileService {
      * @throws RestApiException USER_NOT_FOUND, CREDENTIAL_NOT_FOUND 등
      */
     @Transactional
-    public void updatePassword(Long userId, String newPassword) {
+    public void updatePassword(Integer userId, String newPassword) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RestApiException(UserErrorStatus.USER_NOT_FOUND));
         UserCredential credential = userCredentialRepository.findByUser(user)
